+++ txnbuild/CHANGELOG.md
@@ -8,6 +8,7 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 * In addition to account responses from horizon, transactions and operations can now be built with txnbuild.SimpleAccount structs constructed locally. 
 * Added `MaxTrustlineLimit` which represents the maximum value for a trustline limit.
 * ChangeTrust operation with no `Limit` field set now defaults to `MaxTrustlineLimit`.
+* Add support for building `ManageBuyOffer` operation.
 
 ## [v1.1.0](https://github.com/stellar/go/releases/tag/horizonclient-v1.1.0) - 2019-02-02